moved activity tracker file
