+++ packages/web-sdk/src/instrumentations/_internal/activityWindowTracker.ts
